==> data_cleaning/3_clean_duplicates.ipynb <==
import pandas as pd
df_national =  pd.read_csv('./output/national.csv')
df_national.shape

## sampled data before removing duplicates

df_sample = pd.read_csv("./output/sample_final_full.csv")
df_sample[['sample_id', 'us_state']].groupby(by="us_state").count().reset_index().sort_values(by='sample_id', ascending=False)

## sampled data after removing duplicates

df_sample = pd.read_csv("./output/sample_final_full.csv")
df_sample.Chinese_Name = df_sample.Chinese_Name.str.replace('\n', '').str.replace('\t', '').str.strip()
df_sample.English_Name = df_sample.English_Name.str.replace('\n', '').str.replace('\t', '').str.strip()
print("Sample Size", df_sample.shape)
print("English Unique Names", df_sample.English_Name.unique().shape)
print("Chinese Unique Names", df_sample.Chinese_Name.unique().shape)
print("English+Chinese Unique Names", df_sample[['English_Name', 'Chinese_Name']].drop_duplicates().shape)

# shuffle the order of rows before removing the duplicates
df_sample = df_sample.sample(frac=1)

df_sample = df_sample[~df_sample[['English_Name', 'Chinese_Name']].duplicated(keep="first")]
print(df_sample.shape)
df_sample.to_csv("./output/sample_final_deduplicated_full.csv", index=False)

df_sample.sample(3)

print(df_sample.national_id.unique().shape)
print(df_sample.google_id.unique().shape)
print((3166 / df_national.shape[0])*100)
df_sample[df_sample.national_id == 1].iloc[:, :5]

df_sample[['sample_id', 'us_state']].groupby(by="us_state").count().reset_index().sort_values(by='sample_id', ascending=False)

## get a subset of ACRs with parallel names (unique names)

df_sample_parallel = df_sample[~df_sample.Chinese_Name.isna()]
print("Parallel Sample Size", df_sample_parallel.shape)
print("English Unique Names", df_sample_parallel.English_Name.unique().shape)
print("Chinese Unique Names", df_sample_parallel.Chinese_Name.unique().shape)
print("English+Chinese Unique Names", df_sample_parallel[['English_Name', 'Chinese_Name']].drop_duplicates().shape)

df_sample_parallel.to_csv("./output/sample_final_parallel.csv", index=False)
df_sample_parallel.shape

## merge parallel data with annotated data points

# sanity check if the names match
selectedNUM = 3399

df_ = pd.read_csv("utilities/labeled_datasets/en_golden_label.csv")
print(df_.national_id.unique())
print(df_[df_.national_id == selectedNUM][['English_Name', 'Chinese_Name']])
print(df_sample[df_sample.national_id == selectedNUM][['English_Name', 'Chinese_Name']])

def merge(host_file, guest_file, method, column_name, save_file):
    host = pd.read_csv(host_file)[['national_id']]
    guest = pd.read_csv(guest_file)
    host = host.merge(guest, how=method, on=column_name)
    host = host.dropna(subset=['google_id'])
    host.to_csv(save_file, index=False, encoding='utf-8')
    return host

df_parallel_en = merge('utilities/labeled_datasets/cn_golden_label.csv',
      'output/sample_final_parallel.csv','left','national_id','output/cn_golden_label.csv')
print("English parallel", df_parallel_en.shape)

df_parallel_cn = merge('utilities/labeled_datasets/en_golden_label.csv',
      'output/sample_final_parallel.csv','left','national_id','output/en_golden_label.csv')
print("Chinese parallel", df_parallel_cn.shape)


def split(df_, validation_filename, test_filename):
    df_ = df_.sample(frac=1.0)
    half_size = int(df_.shape[0] / 2)
    validation, test = df_[:half_size], df_[half_size:]
    validation.to_csv(validation_filename, index=False, encoding='utf-8')
    test.to_csv(test_filename, index=False, encoding='utf-8')
    print(validation.shape, test.shape)
    return validation, test

print("English parallel", df_parallel_en.shape)
df_parallel_en = df_parallel_en.drop_duplicates(subset=["English_Name"])
df_parallel_en_val, df_parallel_en_test = split(df_parallel_en, "./output/validation_en.csv", "./output/test_en.csv")
print("Post splitting -- English parallel", df_parallel_en.shape)

print("Chinese parallel", df_parallel_cn.shape)
df_parallel_cn = df_parallel_cn.drop_duplicates(subset=["Chinese_Name"])
df_parallel_cn_val, df_parallel_cn_test = split(df_parallel_cn, "./output/validation_cn.csv", "./output/test_cn.csv")
print("Post splitting -- Chinese parallel", df_parallel_cn.shape)


## sanity check
def check_intersection(list1, list2):
    return len(set(list1).intersection(list2)), set(list1).intersection(list2)

# no English name intersection for the English split
print(check_intersection(df_parallel_en_val.English_Name.unique(), df_parallel_en_test.English_Name.unique()))

# no Chinese name intersection for the Chinese split
print(check_intersection(df_parallel_cn_val.Chinese_Name.unique(), df_parallel_cn_test.Chinese_Name.unique()))

# df_tmp = pd.DataFrame([['A', None], ['A', 1], ['A', 2], ['A', None]], columns=['x', 'y'])
# df_tmp.drop_duplicates(subset=['x', 'y'], keep="last")
